+++ actions/admin/user/removeadmin.php
@@ -3,7 +3,7 @@
  * Revokes admin privileges from a user.
  */
 
-$guid = get_input('guid');
+$guid = (int) get_input('guid');
 $user = get_user($guid);
 
 if ($guid == elgg_get_logged_in_user_guid()) {
